requirements.txt
django-environ==0.10.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pillow==9.4.0
